#402: clean up warnings (part deux)
I C++11'd the casts, and I also changed most floats to doubles, to eliminate a lot of unnecessary conversions.  floats still remain in some of the DSP code for music processing, but elsewhere it's pretty much all double's now.  I also got rid of many unsigned ints, so as to avoid lots of conversions (I don't think it's likely that bugs will show up here, and there was at least one bug in there that was fixed, where a possibly negative value was being cast to unsigned int).  
So, this is general cleanup.  Tested on Mac OS X (but not Win10 or Linux yet).

+++ SquareDesk-DEV/test123/bass_audio.h
@@ -36,11 +36,11 @@ public:
     double                  Current_Position;
     int                     Stream_Volume;
     int                     Stream_Tempo;
-    float                   Stream_Eq[3];
+    double                    Stream_Eq[3];
     int                     Stream_Pitch;
-//    float                   Stream_Pan;
+//    double                    Stream_Pan;
 //    bool                    Stream_Mono;
-    float                   Stream_BPM;
+    double                    Stream_BPM;
     bool                    bPaused;
 
     // LOOP stuff...
@@ -62,9 +62,9 @@ public:
     //Settings
     void SetVolume(int inVolume);
     void SetTempo(int newTempo);  // 100 = normal, 95 = 5% slower than normal
-    void SetEq(int band, float val);  // band = 0,1,2; val = -15.0 .. 15.0 (float) nominal 0.0
+    void SetEq(int band, double  val);  // band = 0,1,2; val = -15.0 .. 15.0 (double ) nominal 0.0
     void SetPitch(int newPitch);  // in semitones, -5 .. 5
-    void SetPan(float newPan);  // -1.0 .. 0.0 .. 1.0
+    void SetPan(double  newPan);  // -1.0 .. 0.0 .. 1.0
 
     void SetLoop(double fromPoint_sec, double toPoint_sec);  // if fromPoint < 0, then disabled
     void ClearLoop();
@@ -72,8 +72,8 @@ public:
     void SetMono(bool on);
 
     //Stream
-    void songStartDetector(const char *filepath, float *pSongStart, float *pSongEnd);
-    void StreamCreate(const char *filepath, float *pSongStart, float *pSongEnd, double i1, double o1);  // returns start of non-silence (seconds)
+    void songStartDetector(const char *filepath, double  *pSongStart, double  *pSongEnd);
+    void StreamCreate(const char *filepath, double  *pSongStart, double  *pSongEnd, double i1, double o1);  // returns start of non-silence (seconds)
 
     void StreamGetLength(void);
     void StreamSetPosition(double Position);
@@ -93,7 +93,7 @@ public:
     void PlayOrStopSoundEffect(int which, const char *filename, int volume = 100);
     void StopAllSoundEffects();
 
-    void StartVolumeDucking(int duckToPercent, float forSeconds);
+    void StartVolumeDucking(int duckToPercent, double  forSeconds);
     void StopVolumeDucking();
 
     // always asks the engine what the state is (NOT CACHED), then returns one of:
